+++ django/test_project2/test_polar/models.py
@@ -173,7 +173,7 @@ class PolarModel(mongomod.Model):
     objects = mongomod.DjongoManager()
 
     class Meta:
-        db_table = "polar2014"
+        db_table = "polar2019"
         app_label = "test_training2"
         managed = False
