+++ django/test_project2/test_project2/templates/base.html
@@ -11,7 +11,7 @@
     <div class="container">
         <header class="d-flex justify-content-center py-3">
           <ul class="nav nav-pills">
-            <li class="nav-item"><a href={% url "test_training2:testpage" %} class="nav-link active" aria-current="page">Home</a></li>
+            <li class="nav-item"><a href="#" class="nav-link active" aria-current="page">Home</a></li>
             <li class="nav-item"><a href={% url "test_polar:showpolar" %} 
               class="nav-link">Summary</a></li> 
             <li class="nav-item"><a href="#" class="nav-link">Analyze</a></li>
