+++ django/test_project2/test_project2/templates/testpage.html
@@ -6,29 +6,29 @@
 <head>
     <meta charset="utf-8">
     <title>testpage</title>
-
+    
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-
-        <style>
-            #textblok {
-              border: 1px solid #999;
-              padding: 10px;  /* Add padding for better visual appearance */
-              background-color: yellow
-            }
-          </style>
-
-    </head>
+    integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
+    
+    <style>
+        #textblok {
+            border: 1px solid #999;
+            padding: 10px;  /* Add padding for better visual appearance */
+            background-color: yellow
+        }
+    </style>
+    
+</head>
 {% block body_block %}
 <body>
+    <script src={% static  "js/training_test.js" %}></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.min.js"
         integrity="sha384-BBtl+eGJRgqQAUMxJ7pMwbEyER4l1g+O15P+16Ep7Q9Q+zqX6gSbd85u4mG4QzX+"
         crossorigin="anonymous"></script>
     <!-- Content here -->
     <h1>Testpage, try to put an image</h1>
     
-    {% comment %} <img src="{% static "images/Dia1.JPG" %}" id="grafje",width="500" height="350", alt="Oh nooo!"> {% endcomment %}
-    <img src="{% static "/images/knmirotterdam_220911_20.png" %}"  id="grafje", width="500" height="600", alt="Oh nooo!">
+    <img src="{% static "images/knmirotterdam_220911_20.png" %}"  id="grafje", width="500" height="600", alt="Oh nooo!">
 
     
     <div class="container">
@@ -59,15 +59,12 @@
                 </select>
             </div>
             <div class="col-6">
-                {% comment %} <p>{{  }}</p> {% endcomment %}
             </div>
             <div class="col-6">
                 <button id="button" onclick="updatePlaatje()">Select graph</button>
             </div>
             
             <p id="textblok" border="1px solid #999">sss </p>
-            <script src={% static  "/js/training_test.js" %}></script>
-            {% comment %} <script src="{% static "training_test.js" %}"></script> {% endcomment %}
         </div>
     </div>
     </div>
